backend/ml-service/ml/requirements.txt
torch
torchvision
pillow
pdf2image
pymupdf
